[emapp] rename VertexWeightBrush to VertexWeightPainter and make owned
